requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
psycopg2-binary
sqlalchemy
pgvector
pydantic-settings
langchain-text-splitters
langchain-google-genai
pypdf
python-multipart
python-jose[cryptography] 
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]
sentence-transformers